cars/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
boto3==1.26.155
botocore==1.29.155
Django==4.2.2
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap4==23.1
django-bootstrap5==23.3
django-cors-headers==4.1.0
django-storages==1.13.2
djangorestframework==3.14.0
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pydantic==1.10.9
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
whitenoise==6.5.0
